+++ lib/Cache/index.d.ts
@@ -1,2 +1,1 @@
-export { default } from './Cache';
 export * from './Cache';