+++ lib/Event/Event.d.ts
@@ -1,8 +1,7 @@
-import Watch from '../Watch';
-import Cache from '../Cache';
+import { Watch } from '../Watch';
+import { Cache } from '../Cache';
 export declare class Event {
     watchers: Set<Watch | Cache>;
-    activeWatchers: Set<Watch | Cache>;
     activeWatcher: Watch;
     add(target: Watch | Cache): void;
     start(): void;
@@ -11,4 +10,3 @@ export declare class Event {
     update(): void;
 }
 export declare const globalEvent: Event;
-export default Event;