+++ src/Cache/index.ts
@@ -1,3 +1,2 @@
-export {default} from './Cache'
 export * from './Cache'