+++ src/Event/index.ts
@@ -1,2 +1,1 @@
-export {default} from './Event'
 export * from './Event'