+++ src/State/index.ts
@@ -1,2 +1,1 @@
-export {default} from './State'
 export * from './State'