+++ src/Watch/index.ts
@@ -1,2 +1,1 @@
-export {default} from './Watch'
 export * from './Watch'